requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
prometheus-fastapi-instrumentator
stripe
python-dotenv
python-multipart